+++ src/Commands/Events/ExchangeCreated.php
@@ -7,7 +7,7 @@ namespace AvtoDev\AmqpRabbitManager\Commands\Events;
 use Interop\Amqp\AmqpTopic as Exchange;
 use Enqueue\AmqpExt\AmqpContext as Connection;
 
-class ExchangeCreated
+final class ExchangeCreated
 {
     /**
      * @var Connection