+++ tests/Commands/Events/ExchangeCreatingTest.php
@@ -4,18 +4,21 @@ declare(strict_types = 1);
 
 namespace AvtoDev\AmqpRabbitManager\Tests\Commands\Events;
 
+use AvtoDev\AmqpRabbitManager\Commands\Events\ExchangeCreating;
+
 /**
  * @covers \AvtoDev\AmqpRabbitManager\Commands\Events\ExchangeCreating<extended>
  */
 class ExchangeCreatingTest extends AbstractEventTestCase
 {
     /**
-     * @small
-     *
-     * @return void
+     * {@inheritDoc}
      */
-    public function testWip(): void
+    public function testConstructorAndProperties(): void
     {
-        self::markTestIncomplete();
+        $event = new ExchangeCreating($this->connection, $this->exchange);
+
+        $this->assertSame($this->connection, $event->connection);
+        $this->assertSame($this->exchange, $event->exchange);
     }
 }