+++ tests/Commands/Events/QueueCreatedTest.php
@@ -4,18 +4,21 @@ declare(strict_types = 1);
 
 namespace AvtoDev\AmqpRabbitManager\Tests\Commands\Events;
 
+use AvtoDev\AmqpRabbitManager\Commands\Events\QueueCreated;
+
 /**
  * @covers \AvtoDev\AmqpRabbitManager\Commands\Events\QueueCreated<extended>
  */
 class QueueCreatedTest extends AbstractEventTestCase
 {
     /**
-     * @small
-     *
-     * @return void
+     * {@inheritDoc}
      */
-    public function testWip(): void
+    public function testConstructorAndProperties(): void
     {
-        self::markTestIncomplete();
+        $event = new QueueCreated($this->connection, $this->queue);
+
+        $this->assertSame($this->connection, $event->connection);
+        $this->assertSame($this->queue, $event->queue);
     }
 }